refactor : exception 위치 변경
