+++ src/main/kotlin/kr/com/hhp/lectureapiserver/lecture/application/exception/CapacityExceededException.kt
@@ -1,4 +1,4 @@
-package kr.com.hhp.lectureapiserver.lecture.exception
+package kr.com.hhp.lectureapiserver.lecture.application.exception
 
 class CapacityExceededException(message: String): RuntimeException(message) {
 }
